CORDA-2490 Attachment in the directory fails to correctly upload (#4627)
The column 'uploader' was marked as not updatable entity so any update wasn't propagated to the database when the attachment was uploaded again and the uploader field was changed to a trusted one. Also 'saveOrUpdate' removed, the entity is already managed, saveOrUpdate is for detached/new entity.

(cherry picked from commit 9f4c8bcea5bb780eec1c1f663302ed92f2c38733)