Correct extending flows java docs and add note on overriding responders
Add `@InitiatedBy` to the java docs on the responder flow, this is
already shown in the kotlin version.

Add a note on overriding responders, instructing developers to still
include the `@InitiatedBy` annotation on the new responder even though
the configuration setup can make developers think that defining the
override will guarantee the initiator and responder will join up
correctly.