ENT-3266: Do not attempt to overrwite an existing CorDapp jar in tests (#1906)
There's a bug with the ServiceLoader which leaks a file handle to the app jar on shutdown. This causes an issue if a mock node is restarted in Windows. To avoid the problem completely we no longer overwrite any existing jars, as the jar to be copied will be same anyway.

(cherry picked from commit 0038a864817d331d7e48d741770b7acc40af2574)