ENT-3257: Logs proxy configuration information on connect failure (#1910)
* ENT-3257: Logs proxy configuration information on connect failure

Currently when a connection fails, no information is logged regarding whether a proxy is configured, and if so, what its address:port is. This may make it more difficult for enterprise customers to diagnose connection issues.
Thus, an additional logging line was added to log proxy configuration in the event of a failed connection. This includes proxy address and port.

* Reduced changes to one line as per review comments