CORDA-2730: Enforce finance cordapp currency restrictions
Changes CashIssueFlow to verify that the finance workflows cordapp that it is executing in
is configured to support the requested currency. The finance workflows cordapp exposes a
configuration property that stores issuable currencies. We check against this list of
issuable currencies before actually issuing currency.

Note that in the default cause, all currencies are supported. In other words, this only
takes effect if the operator deliberately overrides the list of supported currencies with
a custom list.