CORDA-2602 - Update getting setup guide java details (#5019)
Remove some information that is way to specific for installing the
oracle jdk.

Add mention of zulu openjdk to the docs.

Specify that the docs will go through how to install the oracle jdk
only but provide links to amazon corretto, openjdk and zulu.

Also say that we support intellij 2019 which has been released now.