CORDA-2535 - Fixes to IRS demo (#4995)
* CORDA-2535 Move classes from test src to main

For some reason a load of classes were in the test src rather than the
main src even though they were needed during runtime. These have been
moved back into main so they are compiled for use in the demo app.

* CORDA-2535 Fix missing trade id irs-demo home page

Change `deal.ref` to `deal.common.tradeID`

* CORDA-2535 Highlight in irs-demo that `node` module should not be depended on

The demo relies on `node` unless some larger refactoring is made. This
is being kept as it is, but to discourage developers from doing the
same in their CorDapps, a comment has been added to `build.gradle`
explaining why it is there and that developers should not do the same.