ENT-3544: Standardised notary metrics
Implemented metrics in JPAUniquenessProvider and PersistentUniquenessProvider, so that they match MySQLUniquenessProvider as much as possible.
The following metrics were implemented:
commitTimer, inputStatesMeter, conflictCounter, inputStateHistogram, requestProcessingETA, requestQueueCount, requestQueueStateCount, uniqueTxHashCount
This allows us to have a common set of metrics on all enterprise uniqueness providers, so that all notaries record a basic set of metrics.
Note that two of the metrics variables had to be renamed to avoid name conflicts with existent variables in the uniqueness providers - they were inputStateHistogram
and requestQueueCount.

Updated running-a-notary.rst to document the common metrics as well as the two metrics specific to the MySQL notary.

Addressed review comments - moved uniqueTxHashCount updates to near inputStateHistogram updates.

Removed unnecessary *s in closing tag of comments

Removed reference to Graphite in docs.