CORDA-2909 Build `CURRENT_MAJOR_RELEASE` from `build.gradle` in commons-logging
A template directory and file, `Constants`, has been added to allow
gradle to replace the string `"@corda_release_version@"` with the
current corda version provided by gradle. This output will be copied
into the src directory and used for all builds.

The generated `Constants` file has also been checked in to remove
reliance for developers to first run a gradle build to get their
environment fully working. This file does not need to be updated in
the future since gradle will handle building the modules.