ENT-3657: Node janitor: cap retainForDays to 1 year (#2155)
The node janitor performs deduplicated message cleanup in the node's database on shut down. It removes all records older than X days, which is set in the configuration and defaults to the event horizon value in the network parameters.
However, if the event horizon is set to arbitrarily large values such as Max Int, the cleanup logic fails on some databases.