ENT-3830: NodeConfiguration prevents configuring a notary with HSM
Removed validation logic that prevented a notary from being configured with an HSM. This logic was added as part of CORDA-2218.
Added unit test to ensure that a notary can be configured with an HSM and that the config is parsed appropriately.
This check was likely added due to the difficulties inherent in composite keys and the fact that, if an HA notary service is configured with an HSM, the HSM becomes a single point of failure. Nevertheless, there are workarounds available, such as an HA-HSM.