CORDA-3018 Fix `TestDSL` code to work with a new casting made internally
There is a cast of a `ServiceHub` that needs to become a
`ServiceHubCoreInternal`. This is incompatible with `MockServices`. To
resolve this issue, two steps have been taken.
- `as?` has been used in `WireTransaction` at the point of the cast.
  If the cast fails, it must be coming from a test and therefore `true`
  is returned. A comment has been added to highlight this.
- `TestTransactionDSLInterpreter.services` has been updated to
  implement `ServiceHubCoreInternal`. Functionality has been delegated
  to the `ServiceHub` passed to it with missing functions added inline.

This change also highlighted that the `TestDSL` does not properly handle
storing transactions. In other words, `TransactionStorage` does not
return stored transaction. The `validatedTransactions` property of
`TestTransactionDSLInterpreter.services` has been overridden to access
the internally stored transactions that the `TestDSL` holds.