CORDA-3018 Turn `getAllAttachmentsByCriteria` into a `Stream`
Stream results from `AttachmentStorageInternal.getAllAttachmentsByCriteria`.

No database transaction is set inside the function. Streaming results
requires a transaction to be open until the stream processing is
finished.

`NodeAttachmentTrustCalculator` now takes in a `CordaPersistence`.
The `CordaPersistence` property is made nullable to allow easier
addition of the change without requiring a number of changes to the
test util code. A `require` block checks that `CordaPersistence` is
set when it is accessed. This change should be ok to make since only
internal tests should need to test this class.