CORDA-3193: Updated README.md. Note master branch no longer used
Updated the README.md file to provide more detail on the new branching strategy to be used going forward and why master will not be used.

Note that this is intended to be the final commit in the master branch of this repository. Master will not be used anymore -
instead, development work should be done against the appropriate release branch. Branches are named according to the following pattern:
release/os/{major version}.{minor version}

As of September 2019, all work that was previously targeted for the master branch should be rebased on the release/os/4.3 branch.

Please see the above mentioned README.md file for more detail.