TM-26  distribute tests by test duration over a set of k8s forks
We load a previously known set of test results (from TeamCity),
and parse them for their durations.

We then use a gradle plugin to introduce a global dependency
between all tests such that all test projects are compiled and
then inspected so that we know all of the (junit) annotated tests
for all projects.

We can then look up their durations and use a greedy algorithm to
partition them into buckets of approximately equal duration.

Each partition of tests is then run on a k8s fork.