TM-26  Try distributing known tests, followed by random distribution.
If we know the test name and durations (weights) up front, then use
those to distribute over the forks.  We may have some tests that aren't
run that are 'allocated' to a fork.  We make the assumption that it
this won't 'unbalance' the test distribution significantly.

Finally any unknown tests are distributed using the original
random distribution algorithm, which should allocate them to each fork
with equal probability, but deterministrically.

This might make a difference, otherwise I'll revert the change.